requirements.txt
discord.py
flask
requests
BeautifulSoup4
pygithub
